requirements.txt
flask==3.1.0
flask-cors==5.0.1
PyPDF2==3.0.1
python-docx==1.1.2
spacy==3.7.6
fpdf2==2.8.3
streamlit==1.45.0
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
